Input predefined buttons moved to ModuleInput
