+++ LittleOrionEngine.vcxproj.filters
@@ -349,8 +349,10 @@
     <ClCompile Include="Engine\Actions\EditorActionTranslate.cpp">
       <Filter>Engine\Actions</Filter>
     </ClCompile>
-    <ClCompile Include="Engine\SpacePartition\OLAABBTree.cpp" />
     <ClCompile Include="InputManager.cpp" />
+    <ClCompile Include="Engine\SpacePartition\OLAABBTree.cpp">
+      <Filter>Engine\SpacePartition</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Engine\Component\Component.h">
@@ -581,8 +583,10 @@
     <ClInclude Include="Engine\Actions\EditorActionTranslate.h">
       <Filter>Engine\Actions</Filter>
     </ClInclude>
-    <ClInclude Include="Engine\SpacePartition\OLAABBTree.h" />
     <ClInclude Include="InputManager.h" />
+    <ClInclude Include="Engine\SpacePartition\OLAABBTree.h">
+      <Filter>Engine\SpacePartition</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <Filter Include="Libraries">
